update integer field to float

+++ sales/models/airtime_received.py
@@ -14,7 +14,7 @@ class AirtimeReceived(BaseAppModelMixin):
         on_delete=models.CASCADE,
         limit_choices_to={'category': SalesRep.DATA},
         related_name='airtime_received')
-    amount = models.PositiveIntegerField(null=False, blank=False)
+    amount = models.FloatField(null=False, blank=False)
     is_closed = models.BooleanField(default=False)
 
     def clean(self):
